UPDATE migrate some functions
